Bugfix: (Experimental) Smooth scroll - Line number / minimap (#1541)
When experimental.editor.smoothScroll is enabled, the editor surface would scroll smoothly, but the line number and minimap did not. The line number was particularly jarring and could sometimes detract from the smooth scroll effect.

This moves the spring hooks up to a component above, so that the animated scrollX and scrollY can be used by line numbers / minimap.