refactor(definition): Move definition to the feature model (#2168)
This refactors the 'go-to definition' to fit in the feature / language support model.

It moves to a more robust implementation - using a subscription - that fixes a bug with the underline flickering as the cursor moves. This will 'automatically' call `provideDefinition` as the cursor moves, outside of insert mode, and keep the language support state updated.