feat(exthost): Implement storage APIs ($getValue / $setValue) (#2176)
In addition to configuration updaters in #2175 , there is also a `Memento` API extensions can use to set / get values - this is often to use to cache dialog results (prevent them from showing up every time on startup). This implements an endpoint for those APIs through the extensions feature, tying into the persistence mechanism we currently have.