refactor(monorepo): editor-core-types (#2315)
With #2312 and #2313 - moving in `reason-tree-sitter` and `editor-input` - we can now move `editor-core-types` into Onivim proper.

The reason behind this is to update some of the common types - like `Index.t` and `Range.t` - they are not quite setup in a way where we can fully benefit from the type-system. In particular, we don't have typing around byte position vs UTF-8 character position vs screen column - they are just all ints or `Index.t`s. We should be able to leverage the type-system for this, and have auxiliary types, like `ByteRange.t`, so that we can be explicit in our APIs to document whether a function expects a byte index, character index, etc, and have it enforced by the compiler.