fix(ci): Simple lint check (#2337)
Related to #2333 - this is a quick-and-dirty lint script to help verify inadvert `print_endline`, `prerr_endline`, etc are not introduced accidentally in commits.

Was looking for a more robust, full-fledged lint tool for OCaml / ReasonML - but couldn't find anything. Would be a really nice addition! Hopefully at some point we can replace this tool with a real linter.