refactor(explorer): Use luv for populating file explorer (#2450)
More groundwork on the path to #528 - this is some refactoring to support changing the initial load population strategy for the file explorer, which is blocking moving the explorer to its own self-contained feature.

In addition, this moves the file population story to use libuv - and importantly, not block on File I/O. Previously, we were using `Lwt.main_run` to block and populate the directory info synchronously (in addition, we were also traversing a single layer deep inside each folder, which isn't necessary for the initial population). This is actually one bottleneck on Windows start time, at least on my machine (related to #523 ). Removing this blocking is also important for moving to subscriptions, so that we can have predictable performance and running the subscription can be lightweight.