feat(editor/#1444): Wrapping - Part 2 - Add wrapping state to pixel calculations (#2559)
This brings in more of #1969 - wiring up the `Wrapping` module such that we actually use it for calculating pixel position (although, at the moment - the wrap mode is hardcoded to `none` - so no wrapping yet, just plumbing!)

In addition to the functionality from #1969, this also adds a cache for view line -> buffer line mapping, and vice versa. Without it, the traversal needed to calculate the viewline from a bufferline was a bottleneck in the rendering path.

This cache could be further improved by making it incremental - right now, every buffer update, it recalculates in its entirety, but often this isn't necessary (or can just be tweaked). It would help to have more granular buffer updates to implement this, though.

The next step, after this, is to actually hook up the `viewTokens` API from #1969  and adjust the tokenizer / colorizer paths, and wire up the drawing APIs to use that new API. From there, we can start testing the wrapping via experimental configuration settings. There will be some additional work to ensure that `libvim` is aware of the viewport for commands like `H`/`M`/`L`, `gj`, `gk`, etc