fix(completion/#2235): Prioritize fuzzy match over sort text
This adjusts the completion-item sorting to prefer fuzzy matches vs the extension-host-provided sort text.

Once fuzzy-matching is started, as more characters are typed, the `sortText` was still preferred over matches. This would lead to unintuitive results (like in #2235, the `forEachRange` was matched instead of `range` for the filter `ran`, because the `sortText` for `forEachRange` was `f`, whereas for `range` it was `r` - and that was used preferentially.

Fixes #2235 
![Screen Shot 2020-10-10 at 10 03 37 AM](https://user-images.githubusercontent.com/13532591/95660946-0d7a4100-0ae0-11eb-8529-eec1c3049ed6.png)

The `fzy` integration in #1566 has the potential to improve this further.