feat(editor/#1444): WordWrap - Part 4 - Editor viewTokens API (#2577)
Following from #2555 , #2559 , #2572  - this ports over an additional piece of functionality from #1969 for word wrapping.

This pivots the `Editor` API to provide a list of tokens via an `Editor.viewTokens` for a particular line. Before, we were picking up the entire buffer line, but in the case of word wrap, the entire buffer line isn't as useful - each view line only is concerned with the subset of the buffer line contained in the wrapped view.