fix(editor/#760): Confusing scroll/viewport reconciliation (#2597)
Prior to this change, we relied heavily on `libvim`'s concept of scrolling - letting `libvim` keeping track of a `topline` and `leftcol` (which is what terminal vim use for rendering - figuring out what characters to push to the screen).

However, there are several cases where our dependency on that causes some awkward behavior - an example of this is #760  - since we support 'free scroll' with the mouse (scrolling that doesn't move the mouse cursor), when we go back to reconcile that mouse position, `libvim` doesn't actually know about the free-scroll, and there is some awkwardness at that time - the screen doesn't necessarily move to where the user would expect, it moves to the last `topline` that `libvim` knew about.

Also, since word-wrap is managed by our rendering layer - `libvim` also doesn't know about that. So this change - to manage scrolling fully independently of `topline` and `leftcol` - is important for word wrap (#2578) - this lets us decide exactly how to handle all the scroll gestures (`zz`, `zb`, etc) in our view layer. Enabled by https://github.com/onivim/libvim/pull/213

Fixes #760
Required for #2578