fix(scm): Coalesce GotDecorations and scope getOriginalURI (#2615)
These are some minor performance fixes discovered while investigating #2407 - unfortunately, I don't believe these address the root cause based on the information so far - however, these do impact large repos or repository with lots of submodules.

The first issue is that the way we get decorations - the state of files in the workspace, based on the SCM provider (provides coloring in the explorer) - is very inefficient. The exthost APIs are set up to take a batch of URIs, and then return results per URI - however, we were submitting a request per URI. For large repos, with a lot of changes, this can add significant overhead.

Second, when editing a file, we'd ask _every_ provider for the original - in a single repo, this isn't a problem, but in a workspace with lots of submodules or separate git repos, it could be problematic.