fix(osx/#1161): PATH not set for extension host / vim on startup (#2614)
Related to #1161 and #1548 - there is another bug with `PATH` on OSX. When running Onivim from finder, the Onivim process won't get shell environment variables like `PATH`. 

So you get a very small list of `PATH` variables:
![image](https://user-images.githubusercontent.com/13532591/96906340-810c3e80-144e-11eb-9546-ad26fc0fd6eb.png)

This is problematic for the extension host, as well - some extensions look for binaries in the user's current PATH, and if we're not sending the full terminal PATH, some of these utilities could be missed. Likewise for Vim shell commands (ie, `:!`).

It turns out, this isn't a problem unique to Onivim.... happens to Electron apps too, ie: https://stackoverflow.com/questions/62067127/path-variables-empty-in-electron  So we use a similar strategy to the [`fix-path`](https://github.com/sindresorhus/fix-path/blob/8f12bec72ee4319638a43758b40d61c25427404b/index.js#L9) NPM module to address it.