refactor(search): integrate Fzy (#1566)
* Update to bring in new reasonFuzz version.

* Add basic Fzy bits.

* Ignore type error for now.

Need to clean up the interface/think of a nice way of doing this.
We need to get the original item back somehow. Previously, this worked
as everything was item based, so we just returned tuples of items.

Now its list based we can't do that. Worse, we don't even get the same
length list back, and its sorted, which makes matching up a bigger pain.

For now, lets ignore it and do a slow but functional way, and figure out
the crash first.

* Test out fix.

* Fix merge issue.

* Sort results to get sensible output.

* Format.

* Pick up new version of reasonFuzz.

* Bump to pick up list fixes.

* Pick up further changes to fix freeze.

* Lookup item, rather than scary zipping.

Still assume that the index is always there...I see no reason it wouldn't
be, but still something to check.

* Format code.

* Inline makeResult.

* Add Fzy to ThirdPartyLicenses.

* Add new benchmark to run fuzzy matching to completion.

* Format code.

* Swap to straight reason-fzy dependency.

* Pickup fix in reason-fzy.

* Bump reason-fzy version.

* Start merging scores based on match score.

* Bump numbers and add TODO comments.

* Update mergeSortedLists to early stop.

* Use early stopping.

* Fix warnings

* Test out not re-using results.

Can potentially look at adding a smarter version of this later if
we need it?

* Keep around filtered items, but not completed work.

* Test not applicable: Completed work is now reset.

Keeping old completed work around keeps around the old highlights
and scores, which isn't ideal.

* Update test.

* Update rest of tests.

* Tidy up comment and update test.

* Add failing test.

* Remove unused arg.

* Bump reason-fzy.

* Formatting to pass hygiene check

Co-authored-by: Bryan Phelps <bryphe@outlook.com>
Co-authored-by: Bryan Phelps <bryphe@outrunlabs.com>