refactor(input): Add parser for key presses (#2650)
More groundwork for supporting Vim-remaps - this makes our key parser available to be used for remaps. Previously, it was tightly coupled to the `Matcher.t` type, but this removes some unused input code (key up handling, scan code handling) and extracts it. This will be used for handling the `to` part of remaps.