feature(input): Initial handling of Vim-bound ex commands (#2695)
This implements an initial implementation of handling bindings like:

```
nnoremap <silent> <Plug>Hello1 :<C-U>call <SID>hello1()<CR>
```

In other words - plugin-provided `<Plug>` namespaced mappings that call into script-local functions. This is common and used, for example, in `vim-surround`.

This implementation is minimal and only handles a small set of cases, but it's a starting point - a full implementation requires the concept of a `typebuf` (the buffer of pending characters), which we'll need eventually to support more general command forms.

It's enough to start issuing commands for plugins like `vim-surround`, although there are more bugs that need to be fixed (ie, implementing `getchar`) that need to be addressed before it is usable.