fix(perf/#2407): Part 1 - Replace smoothscroll timer hook with model-based animation (#2710)
This is the first part of the fix for #2407 - addressing the hanging scroll timers. This moves away from the hook-based timers, which can potentially leave timers hanging (and have a complicated timing dependency), and replaces it with a model/subscription based approach, which should be more reliable. In particular, the subscriptions used here never leave an `interval` hanging - it's always a one-off fire when the spring animation is active.

This uses the `Spring` model from Revery, but wraps it in a component module (`Component_Animation`), and provides a wrapper model and subscriptions to use. It's sort of a baby-step into migrating our animations from hook-based to model-based.

__TODO:__
- [x] Hook up configuration for `smoothScroll` and `ui.animations`

__Next steps:__
- Replace additional problematic hook-based animations (color transition, status bar animation) with these model-based animations, eventually migrating all of these animations to hook-based if these are successful at mitigating #2407.

Related #2407