fix(editor): Fix several bugs with the active-indent renderer (#2772)
__Issue:__ There were several visual glitches with the active-indent renderer (visible in a screenshot posted in #2353)

Issue 1:

The indent guide would overlap some text

![image](https://user-images.githubusercontent.com/13532591/100487457-8ef05780-30bd-11eb-8409-af5abdab4a95.png)

Issue 2:

When word-wrapping is active, the indent guide would go through wrapped lines:
![image](https://user-images.githubusercontent.com/13532591/100487482-bd6e3280-30bd-11eb-9771-19a3006f21cb.png)

When codelens is active, the indent guide would go through the lens:
![image](https://user-images.githubusercontent.com/13532591/100487538-02926480-30be-11eb-8e48-e73bcdb5a69e.png)


__Defect:__

All of these issues stem from two core problems:
1) If the indentation goes to the end of the file, we'd treat the end coordinate as being at 0 pixels - causing us to draw upward instead of downward (issue 1)
2) When drawing the active line, we'd just draw over all the buffer lines, with no context about word wrap or codelens (issue 2 & 3)

__Fix:__

Tweak the algorithm a bit - it was a two-pass algorithm before, where we'd draw the inactive, and then draw the active line over it.

This is still a two pass algorithm, but first, we calculate the indent levels for visible lines, and then draw the computed indent levels - including the active. The per-buffer-line rendering solves all of the issues - we just needed it to know about the active indent level.