fix(editor/#2620): Shift highlights / markers on buffer update (#3205)
There were a few issues with syntax highlights and other markers (diagnostics, document highlights, search highlights) - cases where there would be frames without highlights:

__Case 1: Pressing enter in insert mode__

![2021-03-02 13 41 09](https://user-images.githubusercontent.com/13532591/109719167-1b972f80-7b5d-11eb-92dc-d5270ab63faa.gif)

The gif recording seems to have missed the errant frame - but there is a frame where the syntax highlights are lost in the newly typed line. The codelens also disappears (same issue), and that is visible in the gif.

Interestingly - this case doesn't occur when creating a new line with `o`. This is due to slightly different shapes in the buffer updates (the enter-in-insert-mode case includes the line that enter was pressed on)

__Case 2: Syntax highlights - single line__

This actually demonstrates both an error in shifting the document highlights and syntax highlights:

![2021-03-02 13 43 07](https://user-images.githubusercontent.com/13532591/109719611-bdb71780-7b5d-11eb-9b51-7c5034a220bd.gif)

__Case 3: Lag in diagnostic updates__

Onivim wasn't updating the diagnostics as the user typed - it would send a buffer update to the extension host, and then wait. This meant there was latency in showing the new error position:

![2021-03-02 13 44 46](https://user-images.githubusercontent.com/13532591/109719578-b4c64600-7b5d-11eb-968f-cc48da020506.gif)

This change adds a 'marker update' - basically instructions for how to offset and adjust markers, syntax highlights, in response to buffer changes. This significantly improves the above cases. 

Depends on https://github.com/onivim/oni2/pull/3139 , so that search highlights can be integrated

Fixes #2620 
Fixes #1459